refactor: Other.jsx의 컴퍼넌트 이동
